Revert "Eric/oauth1"
