+++ client/components/App.js
@@ -2,12 +2,13 @@ import Nav from './Nav'
 import Meta from './Meta'
 // import styles from '../../styles/App.module.css';
 
-
+// here we import GlobalState and wrap the entire App in it to provide our context everywhere...
+import GlobalState from '../context/GlobalState.js'
 
 function App({ children }) {
 
   return (
-    <>
+    <GlobalState>
       <Meta />
       <Nav />
       <div>
@@ -15,7 +16,7 @@ function App({ children }) {
               {children}
           </main>
       </div>
-    </>
+    </GlobalState>
   )
 }
