+++ package.json
@@ -51,21 +51,14 @@
     "body-parser": "^1.19.2",
     "camelcase": "^6.3.0",
     "cookie-parser": "^1.4.6",
-    "cookie-session": "^2.0.0",
     "cors": "^2.8.5",
     "crypto-js": "^4.1.1",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "fontsource-roboto": "^4.0.0",
     "graphql": "^16.3.0",
-    "mongoose": "^6.2.5",
     "next": "^12.1.0",
-<<<<<<< HEAD
-    "passport": "^0.5.2",
-    "passport-github2": "^0.1.12",
-=======
     "pascal-case": "^3.1.2",
->>>>>>> dev
     "pg": "^8.7.3",
     "pluralize": "^8.0.0",
     "react": "^17.0.2",
