+++ server/router.js
@@ -20,4 +20,5 @@ router.post('/schema-user',
     }
 )
 
+
 module.exports = router;
